+++ mingling/src/res.rs
@@ -1,9 +1,11 @@
-// Doc Not Optimize
-mod exit_code;
-pub use exit_code::*;
+#[allow(unused_imports)]
+pub use mingling_core::core_res::*;
 
 mod dirs;
 pub use dirs::*;
 
-#[allow(unused_imports)]
-pub use mingling_core::core_res::*;
+mod exit_code;
+pub use exit_code::*;
+
+mod confirmer;
+pub use confirmer::*;